refactor: static 메서드를 util 파일로 분리
